loader partial

+++ trumptweets/trumptweets/tweets/models.py
@@ -1,7 +1,7 @@
 from django.db.models import (
     CharField,
     DateField,
-    IntegerField,
+    BigIntegerField,
     Model,
 )
 
@@ -9,6 +9,6 @@ from django.db.models import (
 class Tweet(Model):
     raw_text = CharField(max_length=512)     # Tweepy.models.status.text
     author_name = CharField(max_length=256)  # Tweepy.models.status.author.name
-    author_id = IntegerField()  # Tweepy.models.status.author.id
-    created_at = DateField()    # Tweepy.models.created_at (a datetime.datetime object)
-    tweet_id = IntegerField()         # Tweepy.models.status.id
+    author_id = BigIntegerField()  # Tweepy.models.status.author.id
+    created_at = DateField()    # Tweepy.models.status.created_at (a datetime.datetime object)
+    tweet_id = BigIntegerField()   # Tweepy.models.status.id
